kész

+++ AbsztraktOsztaly/Rectangle.cs
@@ -15,7 +15,7 @@ namespace Shape
         public virtual double Height { get => height; set => height = value; }
         public virtual double Width { get => width; set => width = value; }
 
-        public Rectangle(bool isHoley, string color, int height, int width) : base(isHoley, color) 
+        public Rectangle(bool isHoley, string color, double height, double width) : base(isHoley, color) 
         {
             this.height = height;
             this.width = width;
